requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
accelerate==1.11.0
    # via -r requirements.in
ai2-olmo-core==2.3.0
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via
    #   datasets
    #   fsspec
aiosignal==1.4.0
    # via aiohttp
antlr4-python3-runtime==4.9.3
    # via omegaconf
anyio==4.11.0
    # via httpx
attrs==25.4.0
    # via aiohttp
bettermap==1.3.1
    # via ai2-olmo-core
boto3==1.40.65
    # via
    #   cached-path
    #   transformers
botocore==1.40.65
    # via
    #   boto3
    #   s3transfer
cached-path==1.8.0
    # via ai2-olmo-core
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   sacremoses
    #   typer-slim
datasets==2.21.0
    # via -r requirements.in
dill==0.3.8
    # via
    #   bettermap
    #   datasets
    #   multiprocess
filelock==3.20.0
    # via
    #   cached-path
    #   datasets
    #   huggingface-hub
    #   torch
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.6.1
    # via
    #   datasets
    #   huggingface-hub
    #   torch
google-api-core==2.28.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.42.1
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.5.0
    # via google-cloud-storage
google-cloud-storage==2.19.0
    # via cached-path
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.71.0
    # via google-api-core
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   huggingface-hub
huggingface-hub==1.0.0rc6
    # via
    #   -r requirements.in
    #   accelerate
    #   cached-path
    #   datasets
    #   tokenizers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-resources==6.5.2
    # via ai2-olmo-core
jinja2==3.1.6
    # via torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
joblib==1.5.2
    # via sacremoses
markdown==3.6
    # via -r requirements.in
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
networkx==3.5
    # via torch
numpy==1.26.4
    # via
    #   accelerate
    #   ai2-olmo-core
    #   datasets
    #   pandas
    #   transformers
nvidia-cublas-cu12==12.8.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.8.90
    # via torch
nvidia-cuda-nvrtc-cu12==12.8.93
    # via torch
nvidia-cuda-runtime-cu12==12.8.90
    # via torch
nvidia-cudnn-cu12==9.10.2.21
    # via torch
nvidia-cufft-cu12==11.3.3.83
    # via torch
nvidia-cufile-cu12==1.13.1.3
    # via torch
nvidia-curand-cu12==10.3.9.90
    # via torch
nvidia-cusolver-cu12==11.7.3.90
    # via torch
nvidia-cusparse-cu12==12.5.8.93
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.7.1
    # via torch
nvidia-nccl-cu12==2.27.5
    # via torch
nvidia-nvjitlink-cu12==12.8.93
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvshmem-cu12==3.3.20
    # via torch
nvidia-nvtx-cu12==12.8.90
    # via torch
omegaconf==2.3.0
    # via ai2-olmo-core
packaging==25.0
    # via
    #   accelerate
    #   ai2-olmo-core
    #   cached-path
    #   datasets
    #   huggingface-hub
    #   pipdeptree
pandas==2.3.3
    # via datasets
pipdeptree==2.20.0
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via google-api-core
protobuf==6.33.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psutil==7.1.3
    # via accelerate
pyarrow==22.0.0
    # via datasets
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   accelerate
    #   datasets
    #   huggingface-hub
    #   omegaconf
regex==2025.11.3
    # via
    #   sacremoses
    #   transformers
requests==2.32.5
    # via
    #   ai2-olmo-core
    #   cached-path
    #   datasets
    #   google-api-core
    #   google-cloud-storage
    #   transformers
rich==13.7.1
    # via
    #   -r requirements.in
    #   ai2-olmo-core
    #   cached-path
rsa==4.9.1
    # via google-auth
s3transfer==0.14.0
    # via boto3
sacremoses==0.1.1
    # via transformers
safetensors==0.4.3
    # via
    #   -r requirements.in
    #   accelerate
    #   ai2-olmo-core
sentencepiece==0.2.1
    # via transformers
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sympy==1.14.0
    # via torch
tokenizers==0.22.1
    # via -r requirements.in
torch==2.9.0
    # via
    #   -r requirements.in
    #   accelerate
    #   ai2-olmo-core
tqdm==4.66.4
    # via
    #   datasets
    #   huggingface-hub
    #   sacremoses
    #   transformers
transformers==2.3.0
    # via -r requirements.in
triton==3.5.0
    # via torch
typer-slim==0.20.0
    # via
    #   -r requirements.in
    #   huggingface-hub
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   huggingface-hub
    #   torch
    #   typer-slim
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   botocore
    #   requests
xxhash==3.6.0
    # via datasets
yarl==1.22.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
